requirements.txt
RPi.GPIO==0.6.3
requests==2.18.4
